+++ R/data-my_cims.R
@@ -1,0 +1,13 @@
+#' @title The Malaysia Chemical Information Management System Inventory
+#' @description Note that the .tsv file needs to be manually downloaded
+#'   from the official website:
+#'   https://cims.dosh.gov.my/
+#'
+#' @format A data frame with 5,074 rows and 3 variables:
+#' \describe{
+#'   \item{cas_no}{(Character) The Chemical Abstract Service (CAS) Registry
+#'     Number}
+#'   \item{chemical_name}{(Character) The substance name}
+#'   \item{iupac_name}{(Character) The IUPAC-conform substance name}
+#' }
+"my_cims"