+++ R/data-xn_spin.R
@@ -1,0 +1,12 @@
+#' @title The Substances in Preparations in Nordic Countries Inventory
+#' @description Note that the .tsv file needs to be manually downloaded
+#'   from the official website:
+#'   http://www.spin2000.net/spinmyphp/
+#'
+#' @format A data frame with 31,255 rows and 2 variables:
+#' \describe{
+#'   \item{cas_no}{(Character) The Chemical Abstract Service (CAS)
+#'     Registry Number of the substance}
+#'   \item{substances}{(Character) The substance name}
+#' }
+"xn_spin"